requirements.txt
line-bot-sdk==2.4.2
transformers==4.27.4
flask==2.2.3
requests==2.28.2
python-dotenv==1.0.0
